+++ src/include/web_client.hpp
@@ -5,7 +5,8 @@
 class WebClient {
  public:
   static bool get(const std::string &url, const std::string &local_path,
-                  bool silent = false, bool fail_on_error = true);
+                  bool silent = false, bool fail_on_error = true,
+                  bool insecure = false);
 
   static bool postFile(const std::string &url, const std::string &file_path,
                        const std::string &auth = std::string(),